refactor(studio-api): split service config internals
