+++ packages/studio/src/api/services/service-config.ts
@@ -1,497 +1,26 @@
-import {
-  chatCompletion,
-  createLLMClient,
-  fetchWithProxy,
-  getAllEndpoints,
-  GLOBAL_ENV_PATH,
-  resolveServiceModelsBaseUrl,
-  resolveServicePreset,
-  resolveServiceProviderFamily,
-  type ProjectConfig,
-} from "@actalk/inkos-core";
-import { readFile, writeFile } from "node:fs/promises";
-import { join } from "node:path";
-import {
-  filterTextChatModels,
-  isCustomServiceId,
-  isTextChatModelId,
-  serviceConfigKey,
-} from "./service-config-models.js";
-
+export type {
+  EnvConfigStatus,
+  EnvConfigSummary,
+  LLMConfigSource,
+  ServiceConfigEntry,
+  ServiceProbeResult,
+} from "./service-config/types.js";
 export {
   filterTextChatModels,
   isCustomServiceId,
   isTextChatModelId,
   serviceConfigKey,
 } from "./service-config-models.js";
-
-export interface ServiceConfigEntry {
-  service: string;
-  name?: string;
-  baseUrl?: string;
-  temperature?: number;
-  apiFormat?: "chat" | "responses";
-  stream?: boolean;
-}
-
-export type LLMConfigSource = "env" | "studio";
-
-export interface EnvConfigSummary {
-  detected: boolean;
-  provider: string | null;
-  baseUrl: string | null;
-  model: string | null;
-  hasApiKey: boolean;
-}
-
-export interface EnvConfigStatus {
-  project: EnvConfigSummary;
-  global: EnvConfigSummary;
-  effectiveSource: "project" | "global" | null;
-  runtimeUsesEnv: false;
-}
-
-export interface ServiceProbeResult {
-  ok: boolean;
-  models: Array<{ id: string; name: string }>;
-  selectedModel?: string;
-  apiFormat?: "chat" | "responses";
-  stream?: boolean;
-  baseUrl?: string;
-  modelsSource?: "api" | "fallback";
-  error?: string;
-}
-
-function normalizeServiceEntry(serviceId: string, value: Record<string, unknown>): ServiceConfigEntry {
-  if (serviceId.startsWith("custom:")) {
-    return {
-      service: "custom",
-      name: decodeURIComponent(serviceId.slice("custom:".length)),
-      ...(typeof value.baseUrl === "string" && value.baseUrl.length > 0 ? { baseUrl: value.baseUrl } : {}),
-      ...(typeof value.temperature === "number" ? { temperature: value.temperature } : {}),
-      ...(value.apiFormat === "chat" || value.apiFormat === "responses" ? { apiFormat: value.apiFormat } : {}),
-      ...(typeof value.stream === "boolean" ? { stream: value.stream } : {}),
-    };
-  }
-
-  if (serviceId === "custom") {
-    return {
-      service: "custom",
-      ...(typeof value.name === "string" && value.name.length > 0 ? { name: value.name } : {}),
-      ...(typeof value.baseUrl === "string" && value.baseUrl.length > 0 ? { baseUrl: value.baseUrl } : {}),
-      ...(typeof value.temperature === "number" ? { temperature: value.temperature } : {}),
-      ...(value.apiFormat === "chat" || value.apiFormat === "responses" ? { apiFormat: value.apiFormat } : {}),
-      ...(typeof value.stream === "boolean" ? { stream: value.stream } : {}),
-    };
-  }
-
-  return {
-    service: serviceId,
-    ...(typeof value.temperature === "number" ? { temperature: value.temperature } : {}),
-    ...(value.apiFormat === "chat" || value.apiFormat === "responses" ? { apiFormat: value.apiFormat } : {}),
-    ...(typeof value.stream === "boolean" ? { stream: value.stream } : {}),
-  };
-}
-
-export function normalizeConfigSource(value: unknown): LLMConfigSource {
-  return value === "studio" ? "studio" : "env";
-}
-
-export function normalizeServiceConfig(raw: unknown): ServiceConfigEntry[] {
-  if (Array.isArray(raw)) {
-    return raw
-      .filter((entry): entry is Record<string, unknown> => Boolean(entry) && typeof entry === "object")
-      .map((entry) => ({
-        service: typeof entry.service === "string" && entry.service.length > 0 ? entry.service : "custom",
-        ...(typeof entry.name === "string" && entry.name.length > 0 ? { name: entry.name } : {}),
-        ...(typeof entry.baseUrl === "string" && entry.baseUrl.length > 0 ? { baseUrl: entry.baseUrl } : {}),
-        ...(typeof entry.temperature === "number" ? { temperature: entry.temperature } : {}),
-        ...(entry.apiFormat === "chat" || entry.apiFormat === "responses" ? { apiFormat: entry.apiFormat } : {}),
-        ...(typeof entry.stream === "boolean" ? { stream: entry.stream } : {}),
-      }));
-  }
-
-  if (raw && typeof raw === "object") {
-    return Object.entries(raw as Record<string, unknown>)
-      .filter(([, value]) => value && typeof value === "object")
-      .map(([serviceId, value]) => normalizeServiceEntry(serviceId, value as Record<string, unknown>));
-  }
-
-  return [];
-}
-
-export function mergeServiceConfig(existing: ServiceConfigEntry[], updates: ServiceConfigEntry[]): ServiceConfigEntry[] {
-  const merged = new Map(existing.map((entry) => [serviceConfigKey(entry), entry]));
-  for (const update of updates) {
-    merged.set(serviceConfigKey(update), update);
-  }
-  return [...merged.values()];
-}
-
-export async function loadRawConfig(root: string): Promise<Record<string, unknown>> {
-  const configPath = join(root, "inkos.json");
-  const raw = await readFile(configPath, "utf-8");
-  return JSON.parse(raw) as Record<string, unknown>;
-}
-
-export async function saveRawConfig(root: string, config: Record<string, unknown>): Promise<void> {
-  await writeFile(join(root, "inkos.json"), JSON.stringify(config, null, 2), "utf-8");
-}
-
-async function readEnvConfigSummary(path: string): Promise<EnvConfigSummary> {
-  try {
-    const raw = await readFile(path, "utf-8");
-    const values = new Map<string, string>();
-
-    for (const line of raw.split(/\r?\n/)) {
-      const trimmed = line.trim();
-      if (!trimmed || trimmed.startsWith("#")) continue;
-      const match = trimmed.match(/^([A-Za-z_][A-Za-z0-9_]*)=(.*)$/);
-      if (!match) continue;
-      const [, key, value] = match;
-      values.set(key, value.trim());
-    }
-
-    const provider = values.get("INKOS_LLM_PROVIDER") ?? null;
-    const baseUrl = values.get("INKOS_LLM_BASE_URL") ?? null;
-    const model = values.get("INKOS_LLM_MODEL") ?? null;
-    const apiKey = values.get("INKOS_LLM_API_KEY") ?? "";
-    const detected = Boolean(provider || baseUrl || model || apiKey);
-
-    return {
-      detected,
-      provider,
-      baseUrl,
-      model,
-      hasApiKey: apiKey.length > 0,
-    };
-  } catch {
-    return {
-      detected: false,
-      provider: null,
-      baseUrl: null,
-      model: null,
-      hasApiKey: false,
-    };
-  }
-}
-
-export async function readEnvConfigStatus(root: string): Promise<EnvConfigStatus> {
-  const project = await readEnvConfigSummary(join(root, ".env"));
-  const global = await readEnvConfigSummary(GLOBAL_ENV_PATH);
-  return {
-    project,
-    global,
-    effectiveSource: project.detected ? "project" : global.detected ? "global" : null,
-    runtimeUsesEnv: false,
-  };
-}
-
-export async function resolveConfiguredServiceBaseUrl(
-  root: string,
-  serviceId: string,
-  inlineBaseUrl?: string,
-): Promise<string | undefined> {
-  if (inlineBaseUrl?.trim()) return inlineBaseUrl.trim();
-
-  if (!isCustomServiceId(serviceId)) {
-    return resolveServicePreset(serviceId)?.baseUrl;
-  }
-
-  try {
-    const config = await loadRawConfig(root);
-    const services = normalizeServiceConfig((config.llm as Record<string, unknown> | undefined)?.services);
-    const matched = services.find((entry) => serviceConfigKey(entry) === serviceId);
-    return matched?.baseUrl;
-  } catch {
-    return undefined;
-  }
-}
-
-export async function resolveConfiguredServiceEntry(root: string, serviceId: string): Promise<ServiceConfigEntry | undefined> {
-  try {
-    const config = await loadRawConfig(root);
-    const services = normalizeServiceConfig((config.llm as Record<string, unknown> | undefined)?.services);
-    return services.find((entry) => serviceConfigKey(entry) === serviceId);
-  } catch {
-    return undefined;
-  }
-}
-
-function buildProbePlans(
-  preferredApiFormat: "chat" | "responses" | undefined,
-  preferredStream: boolean | undefined,
-): Array<{ apiFormat: "chat" | "responses"; stream: boolean }> {
-  const candidates: Array<{ apiFormat: "chat" | "responses"; stream: boolean }> = [];
-  const seen = new Set<string>();
-  const push = (apiFormat: "chat" | "responses", stream: boolean) => {
-    const key = `${apiFormat}:${stream ? "1" : "0"}`;
-    if (seen.has(key)) return;
-    seen.add(key);
-    candidates.push({ apiFormat, stream });
-  };
-
-  if (preferredApiFormat) {
-    push(preferredApiFormat, preferredStream ?? false);
-    push(preferredApiFormat, !(preferredStream ?? false));
-  }
-  const alternate = preferredApiFormat === "responses" ? "chat" : "responses";
-  push(alternate, false);
-  push(alternate, true);
-  push("chat", false);
-  push("chat", true);
-  push("responses", false);
-  push("responses", true);
-  return candidates;
-}
-
-function buildModelCandidates(args: {
-  preferredModel?: string;
-  configModel?: string;
-  envModel?: string | null;
-  discoveredModels: Array<{ id: string; name: string }>;
-  includeGenericFallbacks?: boolean;
-}): string[] {
-  const seen = new Set<string>();
-  const candidates: string[] = [];
-  const push = (value: string | null | undefined) => {
-    if (!value || value.trim().length === 0) return;
-    const id = value.trim();
-    if (seen.has(id)) return;
-    seen.add(id);
-    candidates.push(id);
-  };
-
-  push(args.preferredModel);
-  push(args.configModel);
-  push(args.envModel ?? undefined);
-  for (const model of args.discoveredModels) push(model.id);
-  if (args.includeGenericFallbacks === false) return candidates;
-  push("gpt-5.4");
-  push("gpt-4o");
-  push("claude-sonnet-4-6");
-  push("MiniMax-M2.7");
-  push("kimi-k2.5");
-  return candidates;
-}
-
-function formatServiceProbeError(args: {
-  readonly service: string;
-  readonly label?: string;
-  readonly baseUrl: string;
-  readonly model?: string;
-  readonly apiFormat?: "chat" | "responses";
-  readonly stream?: boolean;
-  readonly error: string;
-}): string {
-  const rawDetail = args.error
-    .replace(/\n\s*\(baseUrl:[\s\S]*?\)$/m, "")
-    .trim();
-  const upstreamDetail = rawDetail.includes("上游详情：")
-    ? rawDetail
-    : "";
-  const context = [
-    `服务商：${args.label ?? args.service}`,
-    `测试模型：${args.model ?? "未确定"}`,
-    `协议：${args.apiFormat === "responses" ? "Responses" : "Chat / Completions"}${typeof args.stream === "boolean" ? `，${args.stream ? "流式" : "非流式"}` : ""}`,
-    `Base URL：${args.baseUrl}`,
-  ].join("\n");
-
-  if (args.service === "google") {
-    return [
-      "Google Gemini 测试连接失败。",
-      context,
-      "",
-      "请优先检查：",
-      "1. API Key 是否来自 Google AI Studio 的 Gemini API key，而不是 OAuth、Vertex AI 或其它 Google 服务凭据。",
-      "2. 该 key 所属项目是否已启用 Gemini API，并且没有被限制到其它 API、来源或服务。",
-      "3. 当前地区/账号是否允许访问 Gemini API。",
-      "4. 如果 key 曾经泄露，请在 AI Studio 重新生成后再保存。",
-      upstreamDetail ? `\n上游返回：${upstreamDetail}` : "",
-    ].filter(Boolean).join("\n");
-  }
-
-  if (args.service === "moonshot" || args.service === "kimiCodingPlan" || args.service === "kimicode") {
-    return [
-      `${args.label ?? args.service} 测试连接失败。`,
-      context,
-      "",
-      "请优先检查模型是否可用，以及 kimi-k2.x 这类模型是否需要 temperature=1。",
-      rawDetail ? `\n上游返回：${rawDetail}` : "",
-    ].filter(Boolean).join("\n");
-  }
-
-  return [
-    `${args.label ?? args.service} 测试连接失败。`,
-    context,
-    "",
-    "请检查 API Key、模型可用性、账号额度，以及协议类型是否匹配该服务商。",
-    rawDetail ? `\n上游返回：${rawDetail}` : "",
-  ].filter(Boolean).join("\n");
-}
-
-async function fetchModelsFromServiceBaseUrl(
-  serviceId: string,
-  baseUrl: string,
-  apiKey: string,
-  proxyUrl?: string,
-): Promise<{ models: Array<{ id: string; name: string }>; error?: string; authFailed?: boolean }> {
-  const endpoint = isCustomServiceId(serviceId)
-    ? undefined
-    : getAllEndpoints().find((ep) => ep.id === serviceId);
-  const modelsBaseUrl = isCustomServiceId(serviceId)
-    ? baseUrl
-    : endpoint?.modelsBaseUrl ?? (endpoint ? baseUrl : resolveServiceModelsBaseUrl(serviceId) ?? baseUrl);
-  const modelsUrl = modelsBaseUrl.replace(/\/$/, "") + "/models";
-  try {
-    const res = await fetchWithProxy(modelsUrl, {
-      headers: { Authorization: `Bearer ${apiKey}` },
-      signal: AbortSignal.timeout(10_000),
-    }, proxyUrl);
-    if (!res.ok) {
-      const body = await res.text().catch(() => "");
-      return {
-        models: [],
-        error: `服务商返回 ${res.status}: ${body.slice(0, 200)}`,
-        authFailed: res.status === 401 || res.status === 403,
-      };
-    }
-    const json = await res.json() as { data?: Array<{ id: string }> };
-    return {
-      models: (json.data ?? []).map((m) => ({ id: m.id, name: m.id })),
-    };
-  } catch (error) {
-    return {
-      models: [],
-      error: error instanceof Error ? error.message : String(error),
-    };
-  }
-}
-
-export async function probeServiceCapabilities(args: {
-  root: string;
-  service: string;
-  apiKey: string;
-  baseUrl: string;
-  preferredApiFormat?: "chat" | "responses";
-  preferredStream?: boolean;
-  preferredModel?: string;
-  proxyUrl?: string;
-}): Promise<ServiceProbeResult> {
-  const rawConfig = await loadRawConfig(args.root).catch(() => ({} as Record<string, unknown>));
-  const llm = (rawConfig.llm as Record<string, unknown> | undefined) ?? {};
-  const envConfig = await readEnvConfigStatus(args.root);
-  const envModel = envConfig.effectiveSource === "project"
-    ? envConfig.project.model
-    : envConfig.effectiveSource === "global"
-      ? envConfig.global.model
-      : null;
-
-  const baseService = isCustomServiceId(args.service) ? "custom" : args.service;
-  const modelsResponse = await fetchModelsFromServiceBaseUrl(baseService, args.baseUrl, args.apiKey, args.proxyUrl);
-  if (modelsResponse.authFailed) {
-    return {
-      ok: false,
-      models: [],
-      error: modelsResponse.error ?? "API Key 无效或无权访问模型列表。",
-    };
-  }
-  const discoveredModels = modelsResponse.models;
-  const endpoint = getAllEndpoints().find((ep) => ep.id === baseService);
-  const preset = resolveServicePreset(baseService);
-  const discoveredFirstModel =
-    discoveredModels.find((model) => isTextChatModelId(model.id))?.id
-    ?? discoveredModels[0]?.id;
-  const serviceFirstModel =
-    discoveredFirstModel
-    ?? endpoint?.checkModel
-    ?? preset?.knownModels?.[0]
-    ?? endpoint?.models.find((model) => model.enabled !== false)?.id;
-  const useDynamicLocalModels = baseService === "ollama";
-  const useEndpointCheckModel = !useDynamicLocalModels
-    && !isCustomServiceId(args.service)
-    && discoveredModels.length === 0
-    && Boolean(endpoint?.checkModel);
-  const configService = typeof llm.service === "string" ? llm.service : undefined;
-  const configModel = !useEndpointCheckModel && configService === args.service
-    ? typeof llm.defaultModel === "string"
-      ? llm.defaultModel
-      : typeof llm.model === "string"
-        ? llm.model
-        : undefined
-    : undefined;
-  const useCustomFallbacks = isCustomServiceId(args.service);
-  const modelCandidates = buildModelCandidates({
-    preferredModel: args.preferredModel ?? serviceFirstModel,
-    configModel,
-    envModel: useCustomFallbacks ? envModel : undefined,
-    discoveredModels: useEndpointCheckModel ? [] : discoveredModels,
-    includeGenericFallbacks: useCustomFallbacks,
-  });
-
-  if (modelCandidates.length === 0) {
-    return {
-      ok: false,
-      models: [],
-      error: "无法自动确定模型，请先填写可用模型或提供支持 /models 的服务端点。",
-    };
-  }
-
-  let lastError = modelsResponse.error ?? "自动探测失败";
-
-  for (const model of modelCandidates) {
-    for (const plan of buildProbePlans(args.preferredApiFormat, args.preferredStream)) {
-      const client = createLLMClient({
-        provider: resolveServiceProviderFamily(baseService) ?? "openai",
-        service: baseService,
-        configSource: "studio",
-        baseUrl: args.baseUrl,
-        apiKey: args.apiKey.trim(),
-        model,
-        temperature: 0.7,
-        maxTokens: 2048,
-        thinkingBudget: 0,
-        proxyUrl: args.proxyUrl,
-        apiFormat: plan.apiFormat,
-        stream: plan.stream,
-      } as ProjectConfig["llm"]);
-
-      try {
-        await chatCompletion(client, model, [{ role: "user", content: "ping" }], { maxTokens: 2048 });
-        const models = discoveredModels.length > 0
-          ? discoveredModels
-          : endpoint?.models
-            .filter((m) => m.enabled !== false)
-            .filter((m) => isTextChatModelId(m.id))
-            .map((m) => ({ id: m.id, name: m.id }))
-            ?? preset?.knownModels?.map((id) => ({ id, name: id }))
-            ?? [{ id: model, name: model }];
-        return {
-          ok: true,
-          models,
-          selectedModel: model,
-          apiFormat: plan.apiFormat,
-          stream: plan.stream,
-          baseUrl: args.baseUrl,
-          modelsSource: discoveredModels.length > 0 ? "api" : "fallback",
-        };
-      } catch (error) {
-        lastError = formatServiceProbeError({
-          service: baseService,
-          label: endpoint?.label ?? preset?.label,
-          baseUrl: args.baseUrl,
-          model,
-          apiFormat: plan.apiFormat,
-          stream: plan.stream,
-          error: error instanceof Error ? error.message : String(error),
-        });
-      }
-    }
-  }
-
-  return {
-    ok: false,
-    models: discoveredModels,
-    error: lastError,
-  };
-}
+export {
+  mergeServiceConfig,
+  normalizeConfigSource,
+  normalizeServiceConfig,
+} from "./service-config/normalize.js";
+export { readEnvConfigStatus } from "./service-config/env.js";
+export {
+  loadRawConfig,
+  resolveConfiguredServiceBaseUrl,
+  resolveConfiguredServiceEntry,
+  saveRawConfig,
+} from "./service-config/project-config.js";
+export { probeServiceCapabilities } from "./service-config/probe.js";
