+++ packages/studio/src/api/services/service-config/model-discovery.ts
@@ -1,0 +1,44 @@
+import {
+  fetchWithProxy,
+  getAllEndpoints,
+  resolveServiceModelsBaseUrl,
+} from "@actalk/inkos-core";
+import { isCustomServiceId } from "../service-config-models.js";
+
+export async function fetchModelsFromServiceBaseUrl(
+  serviceId: string,
+  baseUrl: string,
+  apiKey: string,
+  proxyUrl?: string,
+): Promise<{ models: Array<{ id: string; name: string }>; error?: string; authFailed?: boolean }> {
+  const endpoint = isCustomServiceId(serviceId)
+    ? undefined
+    : getAllEndpoints().find((ep) => ep.id === serviceId);
+  const modelsBaseUrl = isCustomServiceId(serviceId)
+    ? baseUrl
+    : endpoint?.modelsBaseUrl ?? (endpoint ? baseUrl : resolveServiceModelsBaseUrl(serviceId) ?? baseUrl);
+  const modelsUrl = modelsBaseUrl.replace(/\/$/, "") + "/models";
+  try {
+    const res = await fetchWithProxy(modelsUrl, {
+      headers: { Authorization: `Bearer ${apiKey}` },
+      signal: AbortSignal.timeout(10_000),
+    }, proxyUrl);
+    if (!res.ok) {
+      const body = await res.text().catch(() => "");
+      return {
+        models: [],
+        error: `服务商返回 ${res.status}: ${body.slice(0, 200)}`,
+        authFailed: res.status === 401 || res.status === 403,
+      };
+    }
+    const json = await res.json() as { data?: Array<{ id: string }> };
+    return {
+      models: (json.data ?? []).map((m) => ({ id: m.id, name: m.id })),
+    };
+  } catch (error) {
+    return {
+      models: [],
+      error: error instanceof Error ? error.message : String(error),
+    };
+  }
+}
